Move to a two-phase initialization strategy for all objects.
This is the first step toward moving all kernel operations out of the audio
thread. In the long run we can get this down to 1 semaphore wait and a small
number of semaphore signals and, if we extend miniaudio, we might also be able
to eventually merge our thread with that one on some platforms.