Implement a deferred freeing strategy
We have lots of non-optional allocation, which gets paired with non-optional
freeing, and we need that freeing not to happen in the audio thread.
Additionally, consumers of the public API may be in soft realtime scenarios,
i.e. game loops, which need to meet a deadline.  This means that operations that
can block in the audio thread as well as in any thread from the user will cause
problems.  In practice it is not possible to both have a complex library and
fully eliminate all sources of blocking, but we can do better than we have
been.

This work addresses the freeing story. We free in a background thread that
wakes up on a period in low-to-medium memory pressure situations, and free on
the thread requesting the free if the volume of
deferred frees fills the queue. In general we don't need to do this based off
counting bytes because most Synthizer allocations are around the same size and
intentionally on the smaller side.

Accomplished via using STL custom allocators and making alignFree defer.
Anything using free directly will need to use `deferedFree(free, value)` instead.