Convert the remains of filter_print to filter_repl and apply more filter fixes
This is a program that takes a filter design (hardcoded in the file) and lets
you play around with it by entering frequencies. It shows the theoretical
response at that frequency, then computes the real frequency response
by running a filter.

This was then able to knock out a few more off by ones in our filter
implementation, the effect of which is that noise sounds much closer to online
samples of those algorithms.

Foundational work for reverb, which needs 100% guaranteed working filters.