Get rid of `syz_getO`.
Currently object properties have to use a spinlock. This is very bad because
on Linux, which has a default thread scheduler that can randomly put the user's
thread to sleep on us while it's holding the lock for on the order of 100MS.
Unfortunately, so long as we want refcounted pointers, the LatchCell isn't
suitable, nor are most other options here.

basically, this was the classical textbook case of priority inversion.