Lift the power of 2 restriction on MpscRing.
This was to protect against performance degradation, but the command
infrastructure needs more flexibility on tuning this because commands are going
to be very large objects; we probably want around 10000 or 12000 of them, but
they're going to weigh in at 500+ bytes each.  Since everything else to do with
commands is slower than a modulus and the throughput is low, let's take the
hit.