Stop over-aligning buffers
Some research shows that on modern X86 and modern Arm, the penalties of
unaligned loads that don't cross cache lines are sufficient for our case.
Additionally, we don't currently use explicit SIMD so now is the best time to
get rid of this, and malloc on most platforms will give us 16-byte aligned
buffers anyway, meaning that even though we use unaligned SIMD loads in future,
they're not going to cross cache lines.  Simplify this for memory savings and
to make code more obviously correct/safe.