Introduce BlockBufferCache, a source of buffers sized for a block
This will replace the old thread_local approach, and bounds the number of
blocks by the depth of recursion rather than how many functions across the
library decided to use thread_local.  This may seem counterintuitive,
but by keeping the buffers on a stack we are more likely to keep them in the
cache, there's only ever 1 or 2 levels of recursion, and thread_local was
starting to creep through the whole library.

It's pre-warmed with a couple entries, but takes the hit of allocation while
it warms further.  As things continue, we'll want to tune it further.