Refactor StreamingGenerator and GenerationThread
Instead of a giant ad-hoc mess with atomics, use a couple command queues. We
use ConcurrentQueue which is kind of heavyweight, but that's okay: it's low
frequency, and SpscRing currently blocks. Eventually, the library will implement
it's own version of this for the SPSC case, which will make things faster; but
notably GenerationThread is currently MPSC, which may or may not prove useful
eventually.

Preparatory work for giving StreammingGenerator events in #21