Migrate the decoders to unsigned long long
int64_t seemed like a good idea at the time.  size_t would be a good idea, but
if we ever want to use this via FFI then we should use unsigned long long and
it's theoretically possible that someone might want to do more than 2GB of data
on a 32-bit system, so unsigned long long wins.