All objects which might have automation reserve a config parameter in their constructors which must be NULL
The idea here is that at least C, Rust, and Python won't have to change
anything if we use this argument later: those and most other languages can pass
NULL through the FFI even if the underlying type changes.  It's unclear if and/or
when we would need to use this, but it doesn't hurt to have it around to avoid major
version bumps in the C library.

Refs #73