Fix setting buffer playback positionimmediately after changing the buffer, and before initial connection
But at the cost of introducing another oddity around setting position immediately before setting buffers, which hopefully no one will notice.  We may need to irritate everyone by
changing to free functions rather than the property subsystem here because the order is kind of important, or at least making sure
that it resets only when buffers are changed is.