ModPointer: add an asserting interface to replace raw pointers in debug builds
This is opt-in via an extra parameter to createXXXModPointer because some things like the HRTF
actually rely on raw pointers being passed for, e.g., SIMD overloads.  When opted in, a debug build
(NDEBUG not defined) will never dispatch a raw pointer for the returned ModPointer and
every access will assert that it is in bounds.